Minor adjustment to file handling
When creating the csv files for the new data, I unified the naming and removed superfluous information. The figures don't look as nice with the new data and Figure 3D doesn't come out right on my PC but that might be due to me having different package versions installed.
I added the file to extract the figure data from the raw data.